Implemantata modifica e cancellazione nel controller

+++ src/it/tulchiar/autoscuola/AutoscuolaController.java
@@ -11,6 +11,7 @@ import java.time.format.FormatStyle;
 import java.util.ArrayList;
 import java.util.ResourceBundle;
 
+import it.tulchiar.autoscuola.db.DB_common;
 import it.tulchiar.autoscuola.model.Cliente;
 import it.tulchiar.autoscuola.model.Lettera;
 import it.tulchiar.autoscuola.model.Model;
@@ -20,6 +21,7 @@ import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonBar;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.TableColumn;
@@ -41,6 +43,7 @@ public class AutoscuolaController {
 		this.model = model;
 	}
 
+		
 	@FXML // ResourceBundle that was given to the FXMLLoader
     private ResourceBundle resources;
 
@@ -128,15 +131,15 @@ public class AutoscuolaController {
     @FXML // fx:id="txtNote"
     private TextArea txtNote; // Value injected by FXMLLoader
 
-    @FXML // fx:id="btnInserisciModifica"
-    private Button btnInserisciModifica; // Value injected by FXMLLoader
+    @FXML // fx:id="btnSalva"
+    private Button btnSalva; // Value injected by FXMLLoader
 
     @FXML // fx:id="btnAnnulla"
     private Button btnAnnulla; // Value injected by FXMLLoader
 
     @FXML // fx:id="btnCancella"
     private Button btnCancella; // Value injected by FXMLLoader
-    
+
     @FXML
     void doMostraDettagli(MouseEvent event) {
     		ObservableList<Cliente> clientiSelezionati = tblClienti.getSelectionModel().getSelectedItems();
@@ -254,45 +257,56 @@ public class AutoscuolaController {
     }
 
     @FXML
-    void doInserisciModifica(ActionEvent event) {
+    void doSalva(ActionEvent event) {
     		
-    		int id = -1;
+    		int id = -1; // Stato di nuovo
     		
-    		try {
+    		try { // recupero lo stato attuale
     			id = Integer.parseInt(txtId.getText());
 		} catch (NumberFormatException e) {
 			id = -1;				
 		}
     		
-    		Alert a1 = new Alert(AlertType.INFORMATION, "Id: " + id, ButtonType.OK);
-    		a1.showAndWait();
-	    
-    		if(id == -1) { //Aggiungo un nuovo cliente
-		    	String cognome = txtCognome.getText();
-		    	String nome = txtNome.getText();
-		    	String indirizzo = txtIndirizzo.getText();
-		    	String cap = txtCap.getText();
-		    	String localita = txtCap.getText();
-		    	String provincia = txtProvincia.getText();
-		    	String tipoPatente = txtTipoPatente.getText();
-		    	
-		    	LocalDate dataScadenza = null;
-		    	try {
-		    		dataScadenza = LocalDate.parse(txtDataScadenza.getText());
-		    	} catch(DateTimeParseException e) {
-		    		Alert alert = new Alert(AlertType.ERROR);
-		    		alert.setTitle("Formato data errato");
-		    		alert.setHeaderText("Formato data errato");
-		    		alert.setContentText("La data inserita non è nel formato corretto\\n\\n 2017-12-31");
-		    		alert.showAndWait();
-		    		return;
-		    	}
-		    	
-		    	String telefono = txtTelefono.getText();
-		    	String cellulare = txtCellulare.getText();
-		    	String email = txtEmail.getText();
-		    	String note = txtNote.getText();
-		    	
+    		String cognome = txtCognome.getText();
+	    	String nome = txtNome.getText();
+	    	String indirizzo = txtIndirizzo.getText();
+	    	String cap = txtCap.getText();
+	    	String localita = txtCap.getText();
+	    	String provincia = txtProvincia.getText();
+	    	String tipoPatente = txtTipoPatente.getText();
+    		
+	    	LocalDate dataScadenza = null;
+	    	try { //LocalDate.now().format(DB_common.formatter)
+	    		dataScadenza = LocalDate.parse(txtDataScadenza.getText(), DB_common.formatter);
+	    	} catch(DateTimeParseException e) {
+	    		Alert alert = new Alert(AlertType.ERROR, "La data di scadenza inserita non è nel formato corretto\n\n 2017-12-31", ButtonType.OK);
+	    		alert.setTitle("Formato data errato");
+	    		alert.setHeaderText("Formato data errato");
+	    		alert.showAndWait();
+	    		return;
+	    	}
+	    	
+	    	String telefono = txtTelefono.getText();
+	    	String cellulare = txtCellulare.getText();
+	    	String email = txtEmail.getText();
+	    	String note = txtNote.getText();
+	    	
+	    	LocalDate dataInvioLettera = null;
+	    	try {
+//	    		dataInvioLettera = LocalDate.parse(txtDataInvioLettera.getText());
+	    		dataInvioLettera = LocalDate.parse(txtDataInvioLettera.getText(), DB_common.formatter);
+	    	} catch(DateTimeParseException e) {
+	    		Alert alert = new Alert(AlertType.ERROR, "La data di invio lettera inserita non è nel formato corretto\n\n 2017-12-31", ButtonType.OK);
+	    		alert.setTitle("Formato data errato");
+	    		alert.setHeaderText("Formato data errato");
+	    		alert.showAndWait();
+	    		return;
+	    	}
+	    	
+    		
+    		
+    		if(id == -1) { //Se sono nello stato di nuovo - Aggiungo un nuovo cliente
+   	
 		    	Cliente nuovoCliente = new Cliente(id, cognome, nome, indirizzo, cap, localita, provincia, tipoPatente, dataScadenza, telefono, cellulare, email, note, null);
 		    	
 		    	if(model.add(nuovoCliente)) {
@@ -303,11 +317,15 @@ public class AutoscuolaController {
 		    		alert.showAndWait();
 		    		return;
 		    }
-    		} else { //modifico il cliente con id = txtId
-    			Alert alert = new Alert(AlertType.INFORMATION);
+    		} else { //Se sono nello stato di modifica - modifico il cliente con id = txtId
+    			
+    			Cliente cliente = new Cliente(id, cognome, nome, indirizzo, cap, localita, provincia, tipoPatente, dataScadenza, telefono, cellulare, email, note, dataInvioLettera);
+    		    
+    			model.update(cliente);
+    			
+    			Alert alert = new Alert(AlertType.INFORMATION, "Cliente modificato correttamente.", ButtonType.OK);
 	    		alert.setTitle("Modifica cliente");
-	    		alert.setHeaderText("");
-	    		alert.setContentText("Modifico il cliente con id: " + id);
+	    		alert.setHeaderText("Cliente modificato");
 	    		alert.showAndWait();
 	    		return;
     		}
@@ -322,14 +340,26 @@ public class AutoscuolaController {
     void doCancella(ActionEvent event) {
     		int id = Integer.parseInt(txtId.getText());
     		Cliente cliente = new Cliente(id,null,null);
-    		if(model.delete(cliente)) {
-    			resetPannelloAggiungiModifica();
-    			Alert alert = new Alert(AlertType.INFORMATION, "Cancellazione avvenuta con successo.\n "
-    					+ "(in caso di necessità il record è recuperabile)", ButtonType.OK);
-    			alert.showAndWait();
-    			return;
-    		}
     		
+    		Alert alert = new Alert(Alert.AlertType.CONFIRMATION, 
+    				"Sei sicuro di voler cancellare il Cliente selezionato?", 
+    				ButtonType.NO, ButtonType.YES);
+    		alert.setHeaderText("Cancellazione Cliente");
+    		alert.showAndWait().ifPresent(type -> {
+    			System.out.println(type.toString());
+    		        if (type == ButtonType.YES) {
+    		        	System.out.println("Cancello");
+    		        		if(model.delete(cliente)) {
+    		    			resetPannelloAggiungiModifica();
+    		    			Alert alert2 = new Alert(AlertType.INFORMATION, "Cancellazione avvenuta con successo.\n "
+    		    					+ "(in caso di necessità il record è recuperabile)", ButtonType.OK);
+    		    			alert2.setHeaderText("Cliente cancellato!");
+    		    			alert2.showAndWait();
+    		    			return;
+    		        		}
+    		        }
+    		});	
+    	
     }
     
     
@@ -380,7 +410,7 @@ public class AutoscuolaController {
         assert txtEmail != null : "fx:id=\"txtEmail\" was not injected: check your FXML file 'Autoscuola.fxml'.";
         assert txtDataInvioLettera != null : "fx:id=\"txtDataInvioLettera\" was not injected: check your FXML file 'Autoscuola.fxml'.";
         assert txtNote != null : "fx:id=\"txtNote\" was not injected: check your FXML file 'Autoscuola.fxml'.";
-        assert btnInserisciModifica != null : "fx:id=\"btnInserisciModifica\" was not injected: check your FXML file 'Autoscuola.fxml'.";
+        assert btnSalva != null : "fx:id=\"btnSalva\" was not injected: check your FXML file 'Autoscuola.fxml'.";
         assert btnAnnulla != null : "fx:id=\"btnAnnulla\" was not injected: check your FXML file 'Autoscuola.fxml'.";
         assert btnCancella != null : "fx:id=\"btnCancella\" was not injected: check your FXML file 'Autoscuola.fxml'.";
